task-9 Delete comments
