+++ package.json
@@ -13,7 +13,8 @@
     "axios": "^1.7.9",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.0.2"
+    "react-router-dom": "^7.0.2",
+    "react-toastify": "^11.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
