+++ src/api.js
@@ -1,0 +1,25 @@
+import axios from "axios"
+const baseUrl = 'https://mynew-nc-news.onrender.com/api'
+// get all Article
+export const getArticles = ()=>{
+return axios.get(`${baseUrl}/articles`)
+.then(res => res.data.article)
+}
+// get Articles by id
+export const getArticlesById = (article_id) => {
+return axios.get(`${baseUrl}/articles/${article_id}`)
+.then(res => res.data.articles[0])
+}
+// get All Comments
+export const getAllComments = (article_id) => {
+    return axios.get(`${baseUrl}/articles/${article_id}/comments`)
+    .then(res => res.data.comment)
+}
+// Post a comment
+export const postCommentUrl = (article_id, input) =>{
+return axios.post(`${baseUrl}/articles/${article_id}/comments`,input)
+}
+// Delete a comment
+export const deleteCommmentUrl = (comment_id) => {
+    return axios.delete(`${baseUrl}/comments/${comment_id}`)
+}
